requirements.txt
flask>=2.2.0
python-dotenv>=1.0.1
tqdm>=4.0.0
pypdf>=3.8.0
openai>=1.0.0
pinecone>=8.0.0
gunicorn>=21.2.0